feat(minter): protect delete endpoint

+++ src/app/api/minters/[id]/route.ts
@@ -1,5 +1,8 @@
 import deleteMinter from "@/lib/query/minters/deleteMinter"
 import getMinterById from "@/lib/query/minters/getMinterById"
+import isAdmin from "@/lib/utils/auth/isAdmin"
+import isAuthenticated from "@/lib/utils/auth/isAuthenticated"
+import isCurrentUser from "@/lib/utils/auth/isCurrentUser"
 import { StatusCodes } from "http-status-codes"
 
 export async function GET(req: Request, { params }: { params: { id: string } }) {
@@ -10,7 +13,18 @@ export async function GET(req: Request, { params }: { params: { id: string } }) 
 }
 
 export async function DELETE(req: Request, { params }: { params: { id: string } }) {
+  const session = await isAuthenticated()
+
+  if (!session) {
+    return new Response(null, { status: StatusCodes.UNAUTHORIZED })
+  }
+
   const { id } = params
+
+  if (!isCurrentUser(id, session) && !isAdmin(session)) {
+    return new Response(null, { status: StatusCodes.FORBIDDEN })
+  }
+
   await deleteMinter(id)
 
   return new Response(null, { status: StatusCodes.NO_CONTENT })
